requirements/pip.txt
numpy
matplotlib
pandas
bobleesj.utils
openpyxl
click
